backend/requirements.txt
asgiref==3.7.2
blinker==1.8.2
Django==4.1.13
django-mongoengine==0.5.6
django-rest-framework-mongoengine==3.4.1
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.6.1
mongoengine==0.28.2
psycopg2==2.9.9
pymongo==3.12.1
python-dotenv==1.0.1
pytz==2024.1
sqlparse==0.2.4
tzdata==2024.1
